+++ netpyne/network/subconn.py
@@ -296,6 +296,8 @@ def subcellularConn(self, allCellTags, allPopTags):
                         conn['sec'] = newSec
                         conn['loc'] = newLoc
 
+                        print(newSec, newLoc)
+
                         # find grouped conns 
                         if subConnParam.get('groupSynMechs', None) \
                             and len(subConnParam['groupSynMechs']) > 1 \